requirements.txt
# SN98 ForeverMoney Requirements

# Bittensor
bittensor>=6.0.0

# Web framework for miner
flask>=3.0.0
gunicorn>=21.0.0
requests>=2.31.0
tenacity>=8.2.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Web3 for blockchain interaction
web3>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
